Move SteadyTime to core module to time package
This allows SteadyTime to be used in more modules just like NanoTime
